src/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==1.24.4
click==8.1.7
pyyaml==6.0.2
openai==1.37.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
regex==2024.5.15
requests==2.32.3
scipy==1.10.1
sniffio==1.3.1
tenacity==8.5.0
termcolor==2.4.0
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
